Get code to compile temporarily
